feat: modify local signer from remote one; removed private keys from 'contracts/signer'
