+++ backend/src/apis/mod.rs
@@ -1,3 +1,3 @@
 pub mod address_ownership;
-mod csv_parser;
+pub mod csv_parser;
 pub mod round;
